Split atomic repository and OpenADAS parsing/installing methods.
