+++ cherab/atomic/rates/__init__.pxd
@@ -1,6 +1,6 @@
-# Copyright 2016-2022 Euratom
-# Copyright 2016-2022 United Kingdom Atomic Energy Authority
-# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2018 Euratom
+# Copyright 2016-2018 United Kingdom Atomic Energy Authority
+# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -16,6 +16,8 @@
 # See the Licence for the specific language governing permissions and limitations
 # under the Licence.
 
-# for backward compatibility
-from cherab.atomic.rates cimport *
-
+from cherab.openadas.rates.beam cimport *
+from cherab.openadas.rates.cx cimport *
+from cherab.openadas.rates.pec cimport *
+from cherab.openadas.rates.atomic cimport *
+from cherab.openadas.rates.radiated_power cimport *
