+++ cherab/atomic/rates/atomic.pxd
