+++ cherab/atomic/rates/atomic.pyx
