+++ cherab/atomic/rates/beam.pxd
