+++ cherab/atomic/rates/beam.pyx
