+++ cherab/atomic/rates/pec.pyx
