+++ cherab/atomic/rates/radiated_power.pxd
