+++ cherab/atomic/rates/radiated_power.pyx
