+++ cherab/atomic/repository/beam/emission.py
