+++ cherab/atomic/repository/beam/stopping.py
