+++ cherab/atomic/repository/radiated_power.py
