+++ cherab/atomic/repository/wavelength.py
