+++ demos/emission_models/stark_broadening.py
@@ -1,6 +1,6 @@
-# Copyright 2016-2018 Euratom
-# Copyright 2016-2018 United Kingdom Atomic Energy Authority
-# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -29,7 +29,7 @@ from cherab.core import Species, Maxwellian, Plasma, Line
 from cherab.core.atomic.elements import deuterium, nitrogen
 from cherab.core.model import ExcitationLine, RecombinationLine,\
     MultipletLineShape, StarkBroadenedLine
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 from cherab.tools.plasmas import GaussianVolume
 
 
@@ -41,11 +41,11 @@ sigma = 0.25
 world = World()
 
 # create atomic data source
-adas = OpenADAS(permit_extrapolation=True)
+atomic_data = AtomicData(permit_extrapolation=True)
 
 # PLASMA ----------------------------------------------------------------------
 plasma = Plasma(parent=world)
-plasma.atomic_data = adas
+plasma.atomic_data = atomic_data
 plasma.geometry = Sphere(sigma * 5.0)
 plasma.geometry_transform = None
 plasma.integrator = NumericalIntegrator(step=sigma / 5.0)
