+++ demos/openadas/beam_plasma_interaction_rates.py
@@ -1,8 +1,27 @@
+
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+#
+# Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
+# European Commission - subsequent versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/software/page/eupl5
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied.
+#
+# See the Licence for the specific language governing permissions and limitations
+# under the Licence.
+
 import numpy as np
 import matplotlib
 import matplotlib.pyplot as plt
 from cherab.core.atomic import deuterium, carbon
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 
 # Make Latex available in matplotlib figures
 matplotlib.rcParams.update({'font.size': 12})
@@ -10,10 +29,10 @@ matplotlib.rc('text', usetex=True)
 matplotlib.rc('font', **{'family': 'serif', 'serif': ['Computer Modern']})
 
 # initialise the atomic data provider
-adas = OpenADAS(permit_extrapolation=True)
+atomic_data = AtomicData(permit_extrapolation=True)
 
 # Request beam stopping rate and sample at three different electron temperatures
-bms = adas.beam_stopping_rate(deuterium, carbon, 6)
+bms = atomic_data.beam_stopping_rate(deuterium, carbon, 6)
 beam_energies = [10**x for x in np.linspace(np.log10(5000), np.log10(125000), num=512)]
 bms_rates_1 = [bms(x, 1E19, 1) for x in beam_energies]
 bms_rates_2 = [bms(x, 1E19, 100) for x in beam_energies]
@@ -30,7 +49,7 @@ plt.title("Beam Stopping Rates")
 plt.legend()
 
 # Sample the beam population rates
-bmp = adas.beam_population_rate(deuterium, 2, carbon, 6)
+bmp = atomic_data.beam_population_rate(deuterium, 2, carbon, 6)
 bmp_rates_1 = [bmp(x, 1E19, 1) for x in beam_energies]
 bmp_rates_2 = [bmp(x, 1E19, 100) for x in beam_energies]
 bmp_rates_3 = [bmp(x, 1E19, 1000) for x in beam_energies]
@@ -46,7 +65,7 @@ plt.title("Beam Population Rates")
 plt.legend()
 
 # Sample the beam emission rates
-bme = adas.beam_emission_pec(deuterium, deuterium, 1, (3, 2))
+bme = atomic_data.beam_emission_pec(deuterium, deuterium, 1, (3, 2))
 bme_rates_1 = [bme(x, 1E19, 1) for x in beam_energies]
 bme_rates_2 = [bme(x, 1E19, 100) for x in beam_energies]
 bme_rates_3 = [bme(x, 1E19, 1000) for x in beam_energies]
@@ -62,7 +81,7 @@ plt.title("Beam Emission Rates")
 plt.legend()
 
 # Sample the effective CX emission rates
-cxr = adas.beam_cx_pec(deuterium, carbon, 6, (8, 7))
+cxr = atomic_data.beam_cx_pec(deuterium, carbon, 6, (8, 7))
 cxr_n1, cxr_n2 = cxr
 cxr_rate_1 = [cxr[0](x, 100, 1E19, 1, 1) for x in beam_energies]
 cxr_rate_2 = [cxr[1](x, 1000, 1E19, 1, 1) for x in beam_energies]
