+++ demos/plasmas/analytic_plasma.py
@@ -1,7 +1,7 @@
 
-# Copyright 2016-2018 Euratom
-# Copyright 2016-2018 United Kingdom Atomic Energy Authority
-# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -30,7 +30,7 @@ from cherab.core import Species, Maxwellian, Plasma, Line
 from cherab.core.math import sample3d
 from cherab.core.atomic import deuterium
 from cherab.core.model import ExcitationLine, RecombinationLine
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 
 
 class NeutralFunction:
@@ -106,7 +106,7 @@ world = World()
 # plasma creation #
 
 plasma = Plasma(parent=world)
-plasma.atomic_data = OpenADAS(permit_extrapolation=True)
+plasma.atomic_data = AtomicData(permit_extrapolation=True)
 plasma.geometry = Cylinder(3.5, 2.2, transform=translate(0, 0, -1.1))
 plasma.geometry_transform = translate(0, 0, -1.1)
