+++ demos/plasmas/beam_into_slab.py
@@ -1,4 +1,22 @@
 
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+#
+# Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
+# European Commission - subsequent versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/software/page/eupl5
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied.
+#
+# See the Licence for the specific language governing permissions and limitations
+# under the Licence.
+
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -11,7 +29,7 @@ from cherab.core.math import sample3d
 from cherab.core.atomic import hydrogen, deuterium, carbon, Line
 from cherab.core.model import SingleRayAttenuator, BeamCXLine
 from cherab.tools.plasmas.slab import build_slab_plasma
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 
 
 ###############
@@ -20,7 +38,7 @@ from cherab.openadas import OpenADAS
 world = World()
 
 plasma = build_slab_plasma(peak_density=5e19, impurities=[(carbon, 6, 0.005)], parent=world)
-plasma.atomic_data = OpenADAS(permit_extrapolation=True)
+plasma.atomic_data = AtomicData(permit_extrapolation=True)
 
 ####################
 # Visualise Plasma #
@@ -60,7 +78,7 @@ plt.title("Neutral Density profile in x-z plane")
 ###########################
 # Inject beam into plasma #
 
-adas = OpenADAS(permit_extrapolation=True, missing_rates_return_null=True)
+atomic_data = AtomicData(permit_extrapolation=True, missing_rates_return_null=True)
 
 integration_step = 0.0025
 beam_transform = translate(-0.5, 0.0, 0) * rotate_basis(Vector3D(1, 0, 0), Vector3D(0, 0, 1))
@@ -69,7 +87,7 @@ beam_energy = 50000  # keV
 
 beam_full = Beam(parent=world, transform=beam_transform)
 beam_full.plasma = plasma
-beam_full.atomic_data = adas
+beam_full.atomic_data = atomic_data
 beam_full.energy = beam_energy
 beam_full.power = 3e6
 beam_full.element = deuterium
@@ -84,7 +102,7 @@ beam_full.integrator.min_samples = 10
 
 beam_half = Beam(parent=world, transform=beam_transform)
 beam_half.plasma = plasma
-beam_half.atomic_data = adas
+beam_half.atomic_data = atomic_data
 beam_half.energy = beam_energy / 2
 beam_half.power = 3e6
 beam_half.element = deuterium
@@ -99,7 +117,7 @@ beam_half.integrator.min_samples = 10
 
 beam_third = Beam(parent=world, transform=beam_transform)
 beam_third.plasma = plasma
-beam_third.atomic_data = adas
+beam_third.atomic_data = atomic_data
 beam_third.energy = beam_energy / 3
 beam_third.power = 3e6
 beam_third.element = deuterium
