+++ demos/plasmas/ionisation_balance_2d.py
@@ -1,10 +1,29 @@
+
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+#
+# Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
+# European Commission - subsequent versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/software/page/eupl5
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied.
+#
+# See the Licence for the specific language governing permissions and limitations
+# under the Licence.
+
 from collections.abc import Iterable
 
 import matplotlib.pyplot as plt
 import numpy as np
 from raysect.core.math.function.float import Interpolator1DArray, Interpolator2DArray
 from cherab.core.atomic import neon, hydrogen, helium
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 from cherab.tools.equilibrium import example_equilibrium
 from cherab.tools.plasmas.ionisation_balance import (fractional_abundance, equilibrium_map3d_fractional,
                                                      equilibrium_map3d_from_elementdensity,
@@ -78,8 +97,8 @@ def exp_decay(r, lamb, max_val, lim=None):
 
 equilibrium = example_equilibrium()
 # plot_equilibrium(equilibrium, detail=True)
-# load adas atomic database and define elements
-atomic_data = OpenADAS(permit_extrapolation=True)
+# load atomic database and define elements
+atomic_data = AtomicData(permit_extrapolation=True)
 
 element = neon
 element2 = helium
