+++ demos/plasmas/slab_plasma.py
@@ -1,4 +1,22 @@
 
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+#
+# Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
+# European Commission - subsequent versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/software/page/eupl5
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied.
+#
+# See the Licence for the specific language governing permissions and limitations
+# under the Licence.
+
 import numpy as np
 import matplotlib.pyplot as plt
 from raysect.optical import World
@@ -6,13 +24,13 @@ from raysect.optical import World
 from cherab.core.math import sample3d
 from cherab.core.atomic import hydrogen, carbon
 from cherab.tools.plasmas.slab import build_slab_plasma
-from cherab.openadas import OpenADAS
+from cherab.atomic import AtomicData
 
 
 # make a slab plasma
 world = World()
 plasma = build_slab_plasma(peak_density=5e19, impurities=[(carbon, 6, 0.005)], parent=world)
-plasma.atomic_data = OpenADAS(permit_extrapolation=True)
+plasma.atomic_data = AtomicData(permit_extrapolation=True)
 
 ####################
 # Visualise Plasma #
